requirements.txt
dnspython
pyfiglet
requests
rich
python-whois
beautifulsoup4